+++ SirensMoon/FloorLamp.cpp
@@ -14,9 +14,9 @@ FloorLamp::FloorLamp(Game& game, ModeGame& mode, int ID, SquareLight::SquareLigh
 }
 
 void FloorLamp::Debug(int stageNum, Vector2 window_pos, Vector2 camera_pos){
-	DrawBox(_pos.x+_size.x/2 - 5+window_pos.x-camera_pos.x,
-		_pos.y+_size.y/2 - 5 +window_pos.y-camera_pos.y,
-		_pos.x+_size.x/2 + 5 + window_pos.x - camera_pos.x,
-		_pos.y+_size.y/2 + 5 + window_pos.y - camera_pos.y,
+	DrawBox(static_cast<int>(_pos.x + _size.x / 2 - 5 + window_pos.x - camera_pos.x),
+		static_cast<int>(_pos.y+_size.y/2 - 5 +window_pos.y-camera_pos.y),
+		static_cast<int>(_pos.x+_size.x/2 + 5 + window_pos.x - camera_pos.x),
+		static_cast<int>(_pos.y+_size.y/2 + 5 + window_pos.y - camera_pos.y),
 		GetColor(255, 255, 0), 1);
 }
